refactor(mentor): 2차수정
- validate.mjs에서 validation-ui.mjs 파일 분리하기
- updateButton() 파라미터 hasInvalidInput함수도 같이 받기
